requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
pillow>=10.0
redis>=5.0
rq>=2.0
transformers>=4.40
accelerate>=0.30
httpx>=0.27
pytest>=8.0